Pass scope name as a symbol and send to the model
Directly passing the ActiveRecord::Relation to the scope results in
adding redundant conditions to the inverse scope for all preceding query
conditions. Sending the name of the scope to invert upstream, to the
model class itself, retrieves and inverts only the conditions for that
scope, avoiding any redundant conditions in the WHERE NOT clause.